moved a util function to utils.go
