+++ pkg/analyzer/utils.go
@@ -6,6 +6,8 @@ SPDX-License-Identifier: Apache-2.0
 
 package analyzer
 
+import "path/filepath"
+
 func stopProcessing(stopOn1stErr bool, errs []FileProcessingError) bool {
 	for idx := range errs {
 		if errs[idx].IsFatal() || stopOn1stErr && errs[idx].IsSevere() {
@@ -20,4 +22,17 @@ func appendAndLogNewError(errs []FileProcessingError, newErr *FileProcessingErro
 	logError(logger, newErr)
 	errs = append(errs, *newErr)
 	return errs
+}
+
+// returns a file path without its prefix base dir
+func pathWithoutBaseDir(path, baseDir string) string {
+	if path == baseDir { // baseDir is actually a file...
+		return filepath.Base(path) // return just the file name
+	}
+
+	relPath, err := filepath.Rel(baseDir, path)
+	if err != nil {
+		return path
+	}
+	return relPath
 }
